update BBS scripts to reflect recent changes
This update uses z-scored counts as the response, as this allows for much simpler modeling because we don't need to estimate spatially-varying intercepts for all of the hundreds of species. It also makes evaluation easier, as we can directly compute and compare forecasts because the trends we are evaluating on are zero-centred. I also include updated functionalities for processing the data, thanks to Adam's guidance, and some new functions for cross-validation and comparison of forecasts from trait and nontrait models (using multivariate proper scoring rules, as well as RMSE).